+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file is used to list changes made in each version of phpmyadmin.
 
+## 1.0.1
+
+* Updated blowfish hash creation method
+* Added the pmadb LWRP for creating the PMA's databases to each needed node
+
 ## 1.0.0:
 
 * Initial release of phpmyadmin